refactor: Split recyclerview from ArticleListFragment
